refactor: PostValidator 의 글자수 세기 메서드를 StringUtil로 분리 - week3 - 역할 분리를 위함
